repマクロ編集

+++ template.cpp
@@ -1,5 +1,5 @@
 #include <bits/stdc++.h>
-#define rep(i, n) for (int i = 0; i < (n); i++)
+#define rep(i, n) for (ll i = 0; i < (n); i++)
 using namespace std;
 using uint = unsigned int;
 using ll = long long;
